tak now reads values from the command line! No more editing source code and recompiling to login as non guest/or choose whether to connect to playtak or play in terminal!

+++ source/engine/Engine.h
@@ -29,7 +29,7 @@ class Engine
     int evaluatePos(const Position& position);
     int evaluateResult(Result result);
 public:
-    std::string chooseMove(const Position& position, int timeLimitSeconds = 3);
+    std::string chooseMove(const Position& position, double timeLimitSeconds = 3);
     Move chooseMoveNegamax(const Position& position, Move* move, int depth);
     const MoveBuffer chooseMovesNegamax(const Position& position, Move* move, int depth); // Returns all best moves
